06-best-practices/homework/requirements.txt
pytest
python-dotenv
s3fs
